% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaRestaurar.R
\name{areaRestaurar}
\alias{areaRestaurar}
\title{Delimita a área a ser restaurada}
\usage{
areaRestaurar(clipped_use)
}
\arguments{
\item{clipped_use}{Área delimitada pelo buffer referente ao tamanho da APP com os usos do solo.}
}
\value{
a.cons.area Retorna o valor, em hectares, a ser restaurado dentro do buffer de entrada.
}
\description{
Essa função delimita a área a ser restaurada a partir do objeto resultante da aplicação das funções para recorte do uso dentro das APPs, esta função é especifica para os dados de uso do solo disponiveis na FBDS
}
\examples{

exemplo<-gMicro(mapa_MDA,mapa_RMS,mapa_RMD,mapa_NAS,micro,uso)

area<-areaRestaurar(exemplo)

}
