% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaVegetacao.R
\name{areaVegetacao}
\alias{areaVegetacao}
\title{Delimita a área de APP com vegetação nativa}
\usage{
areaVegetacao(clipped_use)
}
\arguments{
\item{clipped_use}{Área delimitada pelo buffer referente ao tamanho da APP com os usos do solo.}
}
\value{
a.veg.area Retorna o valor, em hectares, das áreas com vegetação dentro do buffer de entrada.
}
\description{
Essa função delimita a área a ser restaurada a partir do objeto resultante da aplicação das funções para recorte do uso dentro das APPs, esta função é especifica para os dados de uso do solo disponiveis na FBDS
}
\examples{

exemplo<-gMicro(mapa_MDA,mapa_RMS,mapa_RMD,mapa_NAS,micro,uso)

area<-areaVegetacao(exemplo)

}
