% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixaFBDS.R
\name{baixaFBDS}
\alias{baixaFBDS}
\title{Baixa os dados da FBDS}
\usage{
baixaFBDS(estado, municipio)
}
\arguments{
\item{estado}{O estado do município}

\item{municipio}{o municipio escolhido}
}
\value{
os dados
}
\description{
Essa função baixa os dados de uso do solo e de APP da FBDS
}
\examples{
dado.municipio<-baixaFBDS(estado, municipio)
}
