% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRS_proj.R
\name{CRS_proj}
\alias{CRS_proj}
\title{Função que reprojeta o ShapeFile usando o wkt fornecido de acordo com a atualização para o PROJ6 e GDAL3.}
\usage{
CRS_proj(obj2Reproj, crs_code)
}
\arguments{
\item{obj2Reproj}{Objeto que precisa ser reprojetado}

\item{crs_code}{Código CRS destino.}
}
\value{
Objeto SpatialPolygonsDataFrames reprojetado.
}
\description{
Essa função reprojeta o ShapeFile usando o wkt fornecido de acordo com a atualização para o PROJ6 e GDAL3.
}
\examples{
meu_shp<-CRS_proj(meu_shp, 31982) #reprojetando para SIRGAS2000/UTM 22S

}
