% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gExplode.R
\name{gExplode}
\alias{gExplode}
\title{Função que explode os polígonos em multipart e inclui os holes}
\usage{
gExplode(to.exp)
}
\arguments{
\item{to.exp}{Objeto cujos polígonos estão em multiparte e precisam ser divididos}
}
\value{
Frag SpatialPolygon ou SpatialPolygonsDataFrames com os polígonos separados
}
\description{
Essa função divide os polígonos que são multipartes
}
\examples{
shp.poligonos.single.parts<-gExplode(shp.poligonos.multipart)

}
