% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGrande.R
\name{gGrande}
\alias{gGrande}
\title{Função unifica a hidrográfia, cria os buffers e corta com o uso do solo.}
\usage{
gGrande(mapa_MDA, mapa_RMS, mapa_RMD, mapa_NAS, grande, uso)
}
\arguments{
\item{mapa_MDA}{Mapa da hidrográfia referente as Massas d'água.}

\item{mapa_RMS}{Mapa da hidrográfia referente aos Rios de margem simples.}

\item{mapa_RMD}{Mapa da hidrográfia referente aos Rios de margem dupla.}

\item{mapa_NAS}{Mapa da hidrográfia referente as Nascentes.}

\item{grande}{Conjunto de propriedades caracterizadas como média, resultantes da função separaTamanho.}

\item{uso}{Mapa do uso do solo do local}
}
\value{
Objeto SpatialPolygonsDataFrames referente ao uso do solo dentro do buffer criado
}
\description{
Essa função unifica a hidrografia e cria os buffer referentes as áreas a serem
restauradas em propriedades de tamanho grande, maiores que dez módulos fiscais,
(15 metros em nascentes e 30 nas demais hidrográfias), retornando a intersecção
deste buffer com o uso do solo
}
\examples{
gGrande<-function(mapa_MDA,mapa_RMS,mapa_RMD,mapa_NAS,media,uso)

}
