% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gHoles.R
\name{gHoles}
\alias{gHoles}
\title{Função para identificar os holes nos fragmentos}
\usage{
gHoles(X, Y)
}
\arguments{
\item{X}{Spdf}
}
\value{
Y Spdf
}
\description{
Essa função identifica os buracos (donut shape polygons) nos fragmentos
}
\examples{
holes.shp<-gHoles(shp)

}
