% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMicro.R
\name{gMicro}
\alias{gMicro}
\title{Função unifica a hidrográfia, cria os buffers e corta com o uso do solo para propriedades de tamanho micro}
\usage{
gMicro(mapa_MDA, mapa_RMS, mapa_RMD = NULL, mapa_NAS, micro, uso)
}
\arguments{
\item{mapa_MDA}{Mapa da hidrográfia referente as Massas d'água.}

\item{mapa_RMS}{Mapa da hidrográfia referente aos Rios de margem simples.}

\item{mapa_RMD}{Mapa da hidrográfia referente aos Rios de margem dupla.}

\item{mapa_NAS}{Mapa da hidrográfia referente as Nascentes.}

\item{micro}{Conjunto de propriedades caracterizadas como micro, resultantes da função}

\item{uso}{Mapa do uso do solo do local}
}
\value{
Objeto SpatialPolygonsDataFrames referente ao uso do solo dentro do buffer criado
}
\description{
Essa função unifica a hidrografia e cria os buffer referentes as áreas a serem
restauradas em micro propriedades (15 metros em nascentes e 5 metros nas demais
hidrográfias), retornando a intersecção deste buffer com o uso do solo
}
\examples{
gHidro<-function(mapa_MDA,mapa_RMS,mapa_RMD,mapa_NAS,micro,uso)

}
