% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPeq_2_4.R
\name{gPeq_2_4}
\alias{gPeq_2_4}
\title{Função unifica a hidrográfia, cria os buffers e corta com o uso do solo para propriedades pequenas de 2 a 4.}
\usage{
gPeq_2_4(mapa_MDA, mapa_RMS, mapa_RMD, mapa_NAS, pequeno_2_4, uso)
}
\arguments{
\item{mapa_MDA}{Mapa da hidrográfia referente as Massas d'água.}

\item{mapa_RMS}{Mapa da hidrográfia referente aos Rios de margem simples.}

\item{mapa_RMD}{Mapa da hidrográfia referente aos Rios de margem dupla.}

\item{mapa_NAS}{Mapa da hidrográfia referente as Nascentes.}

\item{pequeno_2_4}{Conjunto de propriedades caracterizadas como entre dois e quatro módulos fiscais, resultantes da função separaTamanho.}

\item{uso}{Mapa do uso do solo do local}
}
\value{
Objeto SpatialPolygonsDataFrames referente ao uso do solo dentro do buffer criado
}
\description{
Essa função unifica a hidrografia e cria os buffer referentes as áreas a serem
restauradas em propriedades pequenas entre dois e quatro módulos fiscais (15 metros em nascentes e 15 metros nas demais
hidrográfias), retornando a intersecção deste buffer com o uso do solo
}
\examples{
gPeq_2_4<-function(mapa_MDA,mapa_RMS,mapa_RMD,mapa_NAS,pequeno_2_4,uso)

}
