% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUniform.R
\name{makeUniform}
\alias{makeUniform}
\title{Organiza os IDs dos poligonos}
\usage{
makeUniform(spdf)
}
\arguments{
\item{spdf}{O spatialpolygonsdataframe original}
}
\value{
A area com os IDs corrigidos newspdf
}
\description{
Essa função evita que os poligonos tenham o mesmo ID
}
\examples{
makeUniform<-function(spdf)
}
