% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUniformID.R
\name{makeUniformID}
\alias{makeUniformID}
\title{Organiza os IDs dos poligonos 2}
\usage{
makeUniformID(spdf)
}
\arguments{
\item{spdf}{O spatialpolygonsdataframe original}
}
\value{
A area com os IDs corrigidos newspdf
}
\description{
Essa função evita que os poligonos tenham o mesmo ID (Corrigir)
}
\examples{
makeUniformID<-function(spdf)
}
