% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePoli.R
\name{mergePoli}
\alias{mergePoli}
\title{Dissolve os poligonos intersectados}
\usage{
mergePoli(buff)
}
\arguments{
\item{buff}{????}
}
\value{
buffdis Os poligonos com as intersecções dissolvidas
}
\description{
Essa função dissolve os poligonos que por acaso estejam intersectados depois dos recortes com o CAR e os buffers
}
\examples{
mergePoli<-function(buff)
}
