% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRestauraR.R
\encoding{UTF-8}
\name{runRestauraR}
\alias{runRestauraR}
\title{Launch the restauraR Shiny Application}
\usage{
runRestauraR(launchBrowser = TRUE, ...)
}
\arguments{
\item{launchBrowser}{Shiny application launch mode. See runApp function (default launchBrowser = TRUE).}

\item{...}{Arguments used in the runApp function.}
}
\description{
Launches a local version of the restauraR application.
}
\keyword{MainFunction}
