% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPrecision.R
\name{setPrecision}
\alias{setPrecision}
\title{Função para correção do erro de topologia após a criação e recorte do buffers}
\usage{
setPrecision(sp)
}
\arguments{
\item{sp}{????}
}
\value{
sp_exploded ????
}
\description{
Essa função ????
}
\examples{
setPrecision<-function(sp)
}
