% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/appServer.R, R/appUI.R
\encoding{UTF-8}
\name{app}
\alias{app}
\alias{appServer}
\alias{sweetAlert}
\alias{header}
\alias{sidebar}
\alias{controlbar}
\alias{body}
\alias{appUI}
\title{Create the restauraRApp User Interface}
\usage{
appServer(input, output, session)

sweetAlert()

header

sidebar

controlbar

body

appUI
}
\arguments{
\item{input}{Input list for the UI.}

\item{output}{Output list for the UI.}

\item{session}{Environment for the UI.}
}
\description{
Functions to define the user interface and server-side program for the restauraRApp.
}
