% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restauraRApp-package.R
\docType{package}
\encoding{UTF-8}
\name{restauraRApp-package}
\alias{restauraRApp-package}
\alias{restauraRApp}
\title{restauraRApp - App of restauraR Package}
\description{
Create restoration solutions to assemble communities from a regional species pool to achieve multiple functional targets based on simulation.
}
\details{
The restauraR package provides an integrative set of tools for creating restoration solutions and assembling communities from a regional species 
pool. This package implements a user-friendly version of the restauraR package using Shiny.
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Debastiani, V. J., Carlucci, M. B., & Cianciaruso, M. V. (2026). Boosting
multifunctionality through adaptive trait-based species addition in ongoing restoration projects. Ecological 
Applications, 36, e70197. https://doi.org/10.1002/eap.70197
}
\author{
Vanderlei J. Debastiani <vanderleidebastiani@yahoo.com.br>, André G. Coutinho <andreganemcoutinho@gmail.com>, Marcos B. Carlucci <carlucci@ufpr.br>, Marcus V. Cianciaruso <cianciaruso@gmail.com>
}
