% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restauraRApp-package.R
\docType{package}
\encoding{UTF-8}
\name{restauraRApp-package}
\alias{restauraRApp-package}
\alias{restauraRApp}
\title{restauraRApp - App of restauraR Package}
\description{
Create restoration solutions to assemble communities from a regional species pool to achieve multiple functional targets based on simulation.
}
\details{
The restauraR package provides an integrative set of tools to create restoration solutions to assemble communities from a regional 
species pool. The framework allows achieving functional targets following multiple approaches: either starting from empty communities or adding species to ongoing 
restoration sites.
}
\references{
Coutinho, A. G., Carlucci, M. B., & Cianciaruso, M. V. (2023). A framework to apply trait-based ecological 
restoration at large scales. Journal of Applied Ecology, 60, 1562–1571. https://doi.org/10.1111/1365-2664.14439

Coutinho, A. G., Nunes, A., Branquinho, C., Carlucci, M. B., & Cianciaruso, M. V. (2024). Natural regeneration 
enhances ecosystem multifunctionality but species addition can increase it during restoration monitoring. Manuscript 
in preparation.
}
\author{
Vanderlei J. Debastiani <vanderleidebastiani@yahoo.com.br>, André G. Coutinho <andreganemcoutinho@gmail.com>, Marcos B. Carlucci <carlucci@ufpr.br>, Marcus V. Cianciaruso <cianciaruso@gmail.com>
}
